requirements.txt
flask-restplus
flask-marshmallow
Flask-SQLAlchemy
Flask-JWT-Extended==4.3.0
marshmallow-sqlalchemy
Werkzeug==0.16.1
Flask==1.1.2
itsdangerous==2.0.1


# sudo apt install python3-venv
# python3 -m venv venv
# source venv/bin/activate
# pip install flask
# export FLASK_APP=app
# export FLASK_ENV=development
# pip3 freeze > requirements.txt